build: Remove numerous SDK modules and contrib files, update core headers, and add a plugin build configuration.

+++ vlc_sdk/include/vlc_httpd.h
@@ -2,6 +2,7 @@
  * vlc_httpd.h: builtin HTTP/RTSP server.
  *****************************************************************************
  * Copyright (C) 2004-2006 VLC authors and VideoLAN
+ * $Id$
  *
  * Authors: Laurent Aimar <fenrir@via.ecp.fr>
  *
@@ -78,7 +79,7 @@ typedef struct
 
 typedef struct httpd_message_t
 {
-    httpd_client_t *cl; /* NULL if not through a connection e vlc internal */
+    httpd_client_t *cl; /* NULL if not throught a connection e vlc internal */
 
     uint8_t i_type;
     uint8_t i_proto;
@@ -99,7 +100,7 @@ typedef struct httpd_message_t
 
     /* body */
     int64_t i_body_offset;
-    size_t  i_body;
+    int     i_body;
     uint8_t *p_body;
 
 } httpd_message_t;
@@ -121,13 +122,13 @@ VLC_API char* httpd_ServerIP( const httpd_client_t *cl, char *, int * );
 
 typedef struct httpd_file_t     httpd_file_t;
 typedef struct httpd_file_sys_t httpd_file_sys_t;
-typedef int (*httpd_file_callback_t)( httpd_file_sys_t *, httpd_file_t *, uint8_t *psz_request, uint8_t **pp_data, size_t *pi_data );
+typedef int (*httpd_file_callback_t)( httpd_file_sys_t *, httpd_file_t *, uint8_t *psz_request, uint8_t **pp_data, int *pi_data );
 VLC_API httpd_file_t * httpd_FileNew( httpd_host_t *, const char *psz_url, const char *psz_mime, const char *psz_user, const char *psz_password, httpd_file_callback_t pf_fill, httpd_file_sys_t * ) VLC_USED;
 VLC_API httpd_file_sys_t * httpd_FileDelete( httpd_file_t * );
 
 
 typedef struct httpd_handler_t  httpd_handler_t;
-typedef int (*httpd_handler_callback_t)( void *, httpd_handler_t *, char *psz_url, uint8_t *psz_request, int i_type, uint8_t *p_in, int i_in, char *psz_remote_addr, char *psz_remote_host, uint8_t **pp_data, size_t *pi_data );
+typedef int (*httpd_handler_callback_t)( void *, httpd_handler_t *, char *psz_url, uint8_t *psz_request, int i_type, uint8_t *p_in, int i_in, char *psz_remote_addr, char *psz_remote_host, uint8_t **pp_data, int *pi_data );
 VLC_API httpd_handler_t * httpd_HandlerNew( httpd_host_t *, const char *psz_url, const char *psz_user, const char *psz_password, httpd_handler_callback_t pf_fill, void * ) VLC_USED;
 VLC_API void * httpd_HandlerDelete( httpd_handler_t * );
